fix: trying to put tokens in clean file
