+++ packages/core/src/style/index.ts
@@ -1,2 +1,3 @@
 export * from "./quickFormTokensDefinition";
 export * from "./modernQuickFormTokens";
+export * from "./tokens";
