+++ packages/core/src/style/quickFormTokensDefinition.ts
@@ -111,74 +111,3 @@ export const defineQuickFormTokens = (...tokens: Array<Partial<QuickFormTokens>>
  */
 
 //export const quickformtokens1 = camelToKebabCase(defaultQuickFormTokens);
-export type QuickFormTokenVars = {
-    [key in keyof QuickFormTokens]: string
-}
-export const quickformtokens: QuickFormTokenVars  = {
-    "white": "var(--white)",
-    "black": "var(--black)",
-    "warning": "var(--warning)",
-    "success": "var(--success)",
-    "info": "var(--info)",
-    "primary": "var(--primary)",
-    "primaryLighter": "var(--primary-lighter)",
-    "secondary": "var(--secondary)",
-    "background": "var(--background)",
-    "onBackground": "var(--on-background)",
-    "onBackgroundDarker100": "var(--on-background-darker100)",
-    "onBackgroundDarker200": "var(--on-background-darker200)",
-    "onBackgroundDarker300": "var(--on-background-darker300)",
-    "onBackgroundDarker400": "var(--on-background-darker400)",
-    "onBackgroundDarker500": "var(--on-background-darker500)",
-    "onBackgroundDarker600": "var(--on-background-darker600)",
-    "onBackgroundDarker700": "var(--on-background-darker700)",
-    "onBackgroundDarker800": "var(--on-background-darker800)",
-    "onBackgroundDarker900": "var(--on-background-darker900)",
-    "onBackgroundLighter": "var(--on-background-lighter)",
-    "surface": "var(--surface)",
-    "onSurface": "var(--on-surface)",
-    "error": "var(--error)",
-    "onError": "var(--on-error)",
-    "onErrorDarker100": "var(--on-error-darker100)",
-    "onErrorDarker200": "var(--on-error-darker200)",
-    "onErrorDarker300": "var(--on-error-darker300)",
-    "onErrorDarker400": "var(--on-error-darker400)",
-    "onErrorDarker500": "var(--on-error-darker500)",
-    "onErrorDarker600": "var(--on-error-darker600)",
-    "onErrorDarker700": "var(--on-error-darker700)",
-    "onErrorDarker800": "var(--on-error-darker800)",
-    "onErrorDarker900": "var(--on-error-darker900)",
-    "onPrimary": "var(--on-primary)",
-    "onSecondary": "var(--on-secondary)",
-    "fontFamily": "var(--font-family)",
-    "headlineFontSize": "var(--headline-font-size)",
-    "subtitleFontSize": "var(--subtitle-font-size)",
-    "paragraphFontSize": "var(--paragraph-font-size)",
-    "paragraphMobileFontSize": "var(--paragraph-mobile-font-size)",
-    "btnFontSize": "var(--btn-font-size)",
-    "btnFontWeight": "var(--btn-font-weight)",
-    "btnEnterKeyTextFontSize": "var(--btn-enter-key-text-font-size)",
-    "multilineTextFontSize": "var(--multiline-text-font-size)",
-    "multilineTextMobileFontSize": "var(--multiline-text-mobile-font-size)",
-    "questionHeadlineFontSize": "var(--question-headline-font-size)",
-    "questionParagraphFontSize": "var(--question-paragraph-font-size)",
-    "questionNumberFontSize": "var(--question-number-font-size)",
-    "questionInputFontSize": "var(--question-input-font-size)",
-    "questionInputGap": "var(--question-input-gap)",
-    "questionPlaceholderColor": "var(--question-placeholder-color)",
-    "questionHeadlineFontWeight": "var(--question-headline-font-weight)",
-    "questionBottomMargin": "var(--question-bottom-margin)",
-    "questionTopMargin": "var(--question-top-margin)",
-    "questionBorderRadius": "var(--question-border-radius)",
-    "questionPadding": "var(--question-padding)",
-    "questionPaddingBottom": "var(--question-padding-bottom)",
-    "gap1": "var(--gap1)",
-    "gap2": "var(--gap2)",
-    "gap4": "var(--gap4)",
-    "disabledOpacity": "var(--disabled-opacity)",
-    "dividerOpacity": "var(--divider-opacity)",
-    "lowEmphasisOpacity": "var(--low-emphasis-opacity)",
-    "mediumEmphasisOpacity": "var(--medium-emphasis-opacity)",
-    "highEmphasisOpacity": "var(--high-emphasis-opacity)",
-    "slideButtonIconSize": "var(--slide-button-icon-size)" 
-}
